+++ lib/goldiloader/association_options.rb
@@ -4,10 +4,23 @@ module Goldiloader
   module AssociationOptions
     extend self
 
-    OPTIONS = [:auto_include, :fully_load]
+    OPTIONS = [:auto_include, :fully_load].freeze
+
+    # This is only used in Rails 5+
+    module AssociationBuilderExtension
+      def self.build(model, reflection)
+        # We have no callbacks to register
+      end
+
+      def self.valid_options
+        OPTIONS
+      end
+    end
 
     def register
-      if ::ActiveRecord::VERSION::MAJOR >= 4
+      if ::ActiveRecord::VERSION::MAJOR >= 5
+        ActiveRecord::Associations::Builder::Association.extensions << AssociationBuilderExtension
+      elsif ::ActiveRecord::VERSION::MAJOR >= 4
         ActiveRecord::Associations::Builder::Association.valid_options.concat(OPTIONS)
       else
         # Each subclass of CollectionAssociation will have its own copy of valid_options so we need